logging_analysis/Web/requirements.txt
Django==3.0.7
numpy==1.20.0
matplotlib==3.1.1
requests==2.22.0
djangorestframework~=3.12.4
mysqlclient==2.0.3
pymysql==1.0.2
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
cffi==1.15.0
